Add displaying popup with book's info
Add active classes for book's card elements, which are the same as
hover pseudo-class.

Use delegation and add event listeners on every books card to catch
clicks on cards and depend of this calls functions for card buttons.

Add displayBookInfo function which get book card and display popup info
with info about this book.